% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_list_models.R
\name{cq_list_models}
\alias{cq_list_models}
\alias{cq_model_information}
\alias{cq_download_model}
\alias{cq_model_folder}
\alias{cq_model_download_folder}
\alias{cq_model_config_json_file}
\alias{cq_fix_bark_config}
\title{List Coqui Models}
\usage{
cq_list_models(convert = TRUE)

cq_model_information()

cq_download_model(model = "tts_models/en/multi-dataset/tortoise-v2")

cq_model_folder()

cq_model_download_folder(model = "tts_models/en/multi-dataset/tortoise-v2")

cq_model_config_json_file(model = "tts_models/en/multi-dataset/tortoise-v2")

cq_fix_bark_config()
}
\arguments{
\item{convert}{Should the output be converted to R rather than give
a Python object back?  Almost surely you should use \code{TRUE} here.}

\item{model}{name of the model, see \link{cq_list_models}}
}
\value{
A character vector
}
\description{
List Coqui Models
}
\examples{
if (is_coqui_installed()) {
   cq_list_models()
}
}
