% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_tts.R
\name{cq_tts}
\alias{cq_tts}
\title{Text to Speech}
\usage{
cq_tts(
  text,
  model = "tts_models/en/ljspeech/fast_pitch",
  path = tempfile(fileext = ".wav"),
  device = NULL,
  ...
)
}
\arguments{
\item{text}{Text to turn into Speech}

\item{model}{Model to use for speech with Coqui}

\item{path}{Path to output \code{wav} file}

\item{device}{Either \code{cuda} or \code{cpu}.  Uses \code{torch.cuda.is_available} to
swtich, or can be explicitly set here.}

\item{...}{additional arguments to send to the function of
\code{tts_to_file} from the Python \code{TTS.api.TTS()} implementation}
}
\value{
A path to the \code{wav} file
}
\description{
Text to Speech
}
