% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_coqui_condaenv.R
\name{create_coqui_condaenv}
\alias{create_coqui_condaenv}
\title{Install Coqui Conda Environment}
\usage{
create_coqui_condaenv(envname = "coqui", packages = "TTS", ...)
}
\arguments{
\item{envname}{The name of, or path to, a conda environment.}

\item{packages}{character vector of packages to install.  ``TTS"`
is automatically added to the vector, passed to \code{\link[reticulate:py_install]{reticulate::py_install()}}}

\item{...}{additional arguments to pass to  \code{\link[reticulate:conda-tools]{reticulate::conda_create()}}.}
}
\value{
Returns the path to the Python binary associated with the
newly-created conda environment.
}
\description{
Install Coqui Conda Environment
}
\examples{
\dontrun{
  if (!is_coqui_installed()) {
    create_coqui_condaenv()
  }
}
}
