% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_coqui.R
\name{install_coqui}
\alias{install_coqui}
\alias{is_coqui_installed}
\title{Install Coqui}
\usage{
install_coqui(packages = NULL, pip = TRUE, ...)

is_coqui_installed()
}
\arguments{
\item{packages}{character vector of packages to install. \code{"coqui-tts"}
is automatically added to the vector, passed to \code{\link[reticulate:py_install]{reticulate::py_install()}}}

\item{pip}{Boolean; use pip for package installation?
Passed to \code{\link[reticulate:py_install]{reticulate::py_install()}}}

\item{...}{additional arguments to pass to  \code{\link[reticulate:py_install]{reticulate::py_install()}}.}
}
\value{
Nothing (\code{invisible(NULL)}).  Command is called for side effects
}
\description{
Install Coqui
}
\examples{
\dontrun{
  if (!is_coqui_installed()) {
    install_coqui()
  }
}
}
